backend/requirements.txt
requests                       
uvicorn
fastapi  
python-multipart
openai
chromadb
python-dotenv
langchain
langchain-experimental
langchain-community
PyPDF2
dashscope
langchain_openai
langchain-chroma
pymupdf
fitz
pydub
pyogg
ipykernel
SpeechRecognition
aiosqlite